seperate test suite for chat and fim
