+++ litellm/main.py
@@ -3822,7 +3822,7 @@ def completion(  # type: ignore # noqa: PLR0915
                 api_key
                 or litellm.api_key
                 or litellm.inception_key
-                or get_secret_str("INCEPTION_API_KEY")
+                or get_secret("INCEPTION_API_KEY")
             )
 
             _response = openai_text_completions.completion(
